# Uses `safetensors` with `diffusers>=0.10`